+++ src/classic/avdtp_source.h
@@ -287,6 +287,15 @@ void avdtp_source_stream_endpoint_request_can_send_now(uint16_t avddp_cid, uint8
 int avdtp_max_media_payload_size(uint16_t avdtp_cid, uint8_t local_seid);
 
 /**
+ * @brief Register media configuration validator. Can reject insuitable configuration or report stream endpoint as currently busy
+ * @note validator has to return AVDTP error codes like: AVDTP_ERROR_CODE_SEP_IN_USE or AVDTP_ERROR_CODE_UNSUPPORTED_CONFIGURATION
+ *       the callback receives the media configuration in the same format as the existing AVDTP_SUBEVENT_SIGNALING_MEDIA_CODEC_SBC_CONFIGURATION
+ *       and similar
+ * @param callback
+ */
+void avdtp_source_register_media_config_validator(uint8_t (*callback)(const avdtp_stream_endpoint_t * stream_endpoint, const uint8_t * event, uint16_t size));
+
+/**
  * @brief De-Init AVDTP Source.
  */
 void avdtp_source_deinit(void);